take comment related things to separate controller and service
