+++ ludos/backend/src/controllers/user.controller.ts
@@ -157,67 +157,4 @@ export class UserController {
   public async getUserInfoById(@Req() req: AuthorizedRequest) {
     return await this.userService.getUserInfo(req.user.id);
   }
-
-  @ApiOperation({ summary: 'Comment on a post' })
-  @ApiOkResponse({
-    description: 'Comment',
-  })
-  @ApiUnauthorizedResponse({
-    description: 'Invalid Credentials',
-  })
-  @ApiBadRequestResponse({
-    description: 'Bad Request',
-  })
-  @HttpCode(200)
-  @ApiBearerAuth()
-  @UseGuards(AuthGuard)
-  @Post('/write-comment')
-  public async writeComment(
-    @Req() req: AuthorizedRequest,
-    @Body() input: WriteCommentDto,
-  ) {
-    await this.userService.writeComment(req.user.id, input);
-  }
-
-  @ApiOperation({ summary: 'Like a comment' })
-  @ApiOkResponse({
-    description: 'Like Comment',
-  })
-  @ApiUnauthorizedResponse({
-    description: 'Invalid Credentials',
-  })
-  @ApiBadRequestResponse({
-    description: 'Bad Request',
-  })
-  @HttpCode(200)
-  @ApiBearerAuth()
-  @UseGuards(AuthGuard)
-  @Post('/like-comment')
-  public async likeComment(
-    @Req() req: AuthorizedRequest,
-    @Body() input: LikeCommentDto,
-  ) {
-    await this.userService.likeComment(req.user.id, input);
-  }
-
-  @ApiOperation({ summary: 'Dislike a comment' })
-  @ApiOkResponse({
-    description: 'Dislike Comment',
-  })
-  @ApiUnauthorizedResponse({
-    description: 'Invalid Credentials',
-  })
-  @ApiBadRequestResponse({
-    description: 'Bad Request',
-  })
-  @HttpCode(200)
-  @ApiBearerAuth()
-  @UseGuards(AuthGuard)
-  @Post('/dislike-comment')
-  public async dislikeComment(
-    @Req() req: AuthorizedRequest,
-    @Body() input: DislikeCommentDto,
-  ) {
-    await this.userService.dislikeComment(req.user.id, input);
-  }
 }
