+++ ludos/backend/src/services/user.service.ts
@@ -20,9 +20,6 @@ import { ChangePasswordResponseDto } from '../dtos/user/response/change-password
 import { LoginResponseDto } from '../dtos/user/response/login-response.dto';
 import { RegisterResponseDto } from '../dtos/user/response/register-response.dto';
 import { EditUserInfoDto } from '../dtos/user/request/edit-info.dto';
-import { WriteCommentDto } from '../dtos/comment/request/write-comment.dto';
-import { LikeCommentDto } from '../dtos/comment/request/like-comment.dto';
-import { DislikeCommentDto } from '../dtos/comment/request/dislike-comment.dto';
 import { Payload } from '../interfaces/user/payload.interface';
 import { ResetPasswordRepository } from '../repositories/reset-password.repository';
 import { UserRepository } from '../repositories/user.repository';
@@ -34,7 +31,6 @@ export class UserService {
   constructor(
     private readonly userRepository: UserRepository,
     private readonly resetPasswordRepository: ResetPasswordRepository,
-    private readonly commentRepository: CommentRepository,
     private readonly jwtService: JwtService,
     private readonly configService: ConfigService,
     ) {}
@@ -206,56 +202,5 @@ export class UserService {
     response.followedGames = user.followedGames;
 
     return response;
-  }
-
-  public async writeComment(userId: string, writeCommentDto: WriteCommentDto) {
-    let user = await this.userRepository.findUserById(userId);
-
-    if (!user) {
-      throw new HttpException(
-        'No user found with this email',
-        HttpStatus.FORBIDDEN,
-      );
-    }
-
-    // check parent id
-    // post / comment / game review
-    // parent id should be the identifier of one of the above
-
-    let comment = {
-      author: userId, //user,
-      text: writeCommentDto.text,
-      parentId: writeCommentDto.parentId,
-      likes: 0,
-      dislikes: 0,
-      timestamp: new Date(),
-    }
-    await this.commentRepository.createComment(comment);
-  }
-
-  public async likeComment(userId: string, likeCommentDto: LikeCommentDto) {
-    let comment = await this.commentRepository.findCommentById(likeCommentDto.commentId);
-
-    if (!comment) {
-      throw new HttpException(
-        'No comment found with this id',
-        HttpStatus.FORBIDDEN,
-      );
-    }
-
-    await this.commentRepository.incrementLikeCount(likeCommentDto.commentId);
-  }
-
-  public async dislikeComment(userId: string, dislikeCommentDto: DislikeCommentDto) {
-    let comment = await this.commentRepository.findCommentById(dislikeCommentDto.commentId);
-
-    if (!comment) {
-      throw new HttpException(
-        'No comment found with this id',
-        HttpStatus.FORBIDDEN,
-      );
-    }
-
-    await this.commentRepository.incrementDislikeCount(dislikeCommentDto.commentId);
   }
 }
